🛡️ Guardian (JULES): Consolidate CodeEditor and DiffView utilities
WHAT:
- Merged src/components/book/elements/code-editor.tsx into src/components/ui/code-editor.tsx
- Merged src/components/book/elements/diff-view.tsx into src/components/ui/diff-view.tsx
- Updated barrel files and standard imports to use the canonical ui components
- Preserved backward compatibility for MDX book rendering
- Deleted obsolete book element copies

WHY:
- Component duplication created inconsistent feature sets (e.g., standard UI CodeEditor missing copy buttons, while MDX versions lacked debounce/Monaco theming).
- Removing duplicates reduces bundle/test surface area and aligns with the single-source-of-truth standard.

JULES CONTEXT:
- Checked .Jules/task-log.md: No Autonomous conflicts.
- Session logged: .Jules/guardian/2026-05-14/session-report.md
- Reference: .Jules/JULES.md

IMPACT:
- Files: 4 → 2 (-50%)
- Code duplication: Removed 2 redundant React component hierarchies.

VERIFICATION:
- All tests pass
- TypeScript check: ✓
- Linting: ✓
- Build: ✓
- JULES compliance: ✓
