+++ .Jules/task-log.md
@@ -1,0 +1,9 @@
+## 2026-05-14 16:55 - [GUARDIAN] Session started
+- Directory: .Jules/guardian/2026-05-14
+- Phase: PRE-FLIGHT
+- JULES Check: COMPLETE
+
+## 2026-05-14 18:29 - [GUARDIAN] Task completed
+- Status: COMPLETED
+- Files: src/components/ui/code-editor.tsx, src/components/ui/diff-view.tsx, src/components/book/elements/code-editor.tsx, src/components/book/elements/diff-view.tsx, src/components/book/elements/index.ts, src/components/book/interactive.tsx
+- Details: Consolidated CodeEditor and DiffView. Deleted redundant files from book/elements. Maintained full backward compatibility for MDX files. Impact: 4 files -> 2 files.
